requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
pillow>=9.0.0
numpy>=1.24.0
datetime